pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "playwright-python-boilerplate"
version = "1.0.0"
description = "API Testing Framework with Playwright"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pytest>=7.4.0",
    "playwright>=1.40.0",
    "pytest-playwright>=0.4.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "Faker>=20.0.0",
    "requests>=2.31.0",
    "structlog>=23.0.0",
    "pytest-html>=4.0.0",
    "allure-pytest>=2.13.0",
    "allure-python-commons>=2.13.0",
    "tenacity>=8.2.0",
    "httpx>=0.25.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.11.0",
    "isort>=5.10.1",
    "pre-commit>=2.20.0",
    "playwright>=1.40.0",
    "flake8>=6.0.0",
    "mypy>=1.9.0",
]

[tool.black]
line-length = 119

[tool.ruff]
line-length = 119
extend-select = ["E", "F", "W", "T" ]

[tool.flake8]
max-line-length = 119

[tool.isort]
line_length = 119

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
markers = [
    "smoke: Smoke tests",
    "regression: Regression tests",
]
